requirements.txt
python-telegram-bot
instaloader
spotipy
requests
yt-dlp
